# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Stellar wallet backend — auth, encrypted credential backup, and account recovery.",
        "title": "Latch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/accounts": {
            "get": {
                "description": "Returns every smart account (seed and passkey wallets) owned by the session user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "List the session user's smart accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/accounts/set-active": {
            "post": {
                "description": "Sets a client-readable cookie recording which smart account is active. Purely a cookie write — no server-side persistence.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Set the active smart account cookie",
                "parameters": [
                    {
                        "description": "Smart account address to mark active",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.setActiveAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/counter": {
            "get": {
                "description": "Simulates the demo counter contract's get() function on testnet. Returns 0 if the simulation doesn't return a success result (e.g. an unset counter).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "counter"
                ],
                "summary": "Get the demo counter value",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.counterResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/multisig/accounts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-accounts"
                ],
                "summary": "List the session user's deployed multisig accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/multisig/drafts": {
            "get": {
                "description": "Returns the session user's in-progress (collecting) draft, or {\"draft\":null} if none exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-drafts"
                ],
                "summary": "Get the session user's active multisig draft",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Must be 1",
                        "name": "active",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Starts a new multisig account draft owned by the session user, generating an invite token other members join with.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "multisig-drafts"
                ],
                "summary": "Create a new multisig draft",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/on-ramp/pool": {
            "get": {
                "description": "Dev-only (403 in production). Returns the on-ramp pool account's XLM balance and up to 20 recent transactions, optionally filtered to a single memo.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "on-ramp"
                ],
                "summary": "Get the on-ramp pool account snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter recent transactions to this memo (e.g. an intent's memoId)",
                        "name": "memo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.onRampPoolResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/sign-payload": {
            "post": {
                "description": "Stores an unsigned transaction XDR plus a callback URL under a single-use, TTL-bounded reference (60s-3600s, default 600s). The referenced payload is deleted-on-read: GET consumes it exactly once. callback must be https://, or http://localhost / http://127.0.0.1 for local development.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sign-payload"
                ],
                "summary": "Store a sign payload for out-of-band signing",
                "parameters": [
                    {
                        "description": "Sign payload to store",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.createSignPayloadRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.createSignPayloadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_webapp.counterResponse": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
        "internal_handler_webapp.createSignPayloadResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string",
                    "example": "2026-07-02T12:10:00Z"
                },
                "payloadRef": {
                    "type": "string",
                    "example": "sp_1a2b3c4d5e6f7890abcdef1234567890"
                }
            }
        },
...
        "internal_handler_webapp.setActiveAccountRequest": {
            "type": "object",
            "required": [
                "smartAccountAddress"
            ],
            "properties": {
                "smartAccountAddress": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_webapp.webappErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "internal_error"
                },
                "error": {
                    "type": "string",
                    "example": "internal error"
                },
                "message": {
                    "type": "string",
                    "example": "internal error"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RecoveryAuth": {
            "description": "Enter: Bearer \u003crecovery_token\u003e  (obtained from POST /v1/recovery/verify)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}